move param_groups to resnet_fpn_utils
