+++ icevision/backbones/resnet_fpn.py
@@ -8,12 +8,11 @@ __all__ = [
     "resnext101_32x8d_fpn",
     "wide_resnet50_2_fpn",
     "wide_resnet101_2_fpn",
-    "param_groups",
-    "patch_param_groups",
 ]
 
 from icevision.imports import *
 from icevision.utils import *
+from icevision.backbones.resnet_fpn_utils import *
 from torchvision.models.detection.backbone_utils import resnet_fpn_backbone
 
 
@@ -58,21 +57,3 @@ def wide_resnet50_2_fpn(pretrained: bool = False, **kwargs):
 
 def wide_resnet101_2_fpn(pretrained: bool = False, **kwargs):
     return _resnet_fpn("wide_resnet101_2", pretrained=pretrained, **kwargs)
-
-
-def param_groups(model: nn.Module) -> List[nn.Parameter]:
-    body = model.body
-
-    layers = []
-    layers += [nn.Sequential(body.conv1, body.bn1)]
-    layers += [getattr(body, l) for l in list(body) if l.startswith("layer")]
-    layers += [model.fpn]
-
-    _param_groups = [list(layer.parameters()) for layer in layers]
-    check_all_model_params_in_groups2(model, _param_groups)
-
-    return _param_groups
-
-
-def patch_param_groups(model: nn.Module) -> None:
-    model.param_groups = MethodType(param_groups, model)
